refactor(core): use consistent import style for the logger
1. move it into utils
2. do not mix default and named exports
3. adjust usage to new location + only named export
